Reuse Basix element instead of recreating from generated code (#3136)
* remove create_functionspace from UFCx functionspace

* Create basix elements in demos instead of using generated code

* update matrix test

* <double>

* update test

* wrap FiniteElement(basix_element, block_size) initialiser

* create function space from element where possible

* remove failicg check of signatures

* dtype

* remove duplicate dtype

* Basix branch

* update tests

* ruff

* ruff format with updated ruff

* update test

* Make mixed elements work without FFCx

* ruff

* typing

* isort

* brnach

* petsc realtype?

* ruff

* ScalarType not RealType

* real?

* simplify?

* dtypes

* dtype eps

* quadrature element without ufcx

* working on removing ufcx_dofmap

* fix tests

* simplifying

* use sub_e not ufcx_dofmap

* goodbye ufcx_dofmap

* simplify

* goodbye ufcx_dofmap

* remove old functions that use ufcx_dofmap

* remove ufcx_dofmap from docs, ruff

* comment out last ufcx_element

* doc

* remove jitting of elements

* Update cpp/dolfinx/fem/FiniteElement.cpp


* Update cpp/dolfinx/fem/FiniteElement.h


* Use constructor initialisers

* Interpolation for symmetric tensor

* add test

* correct test

* ffc branch

* main ffcx

* tidy test

* doc

* dtype

* remove dtype

* remove dtype from test

* symmetric

* ruff

* Make interpolation work on more than one cell

* use ordering of lower part not upper part

* make test ues 3 by 3 not 6 by 6

* move interpolation test

* Fix #3164

* is_symmetric

* , False

* use Basix main

* remove basix.default_scalar_type

* dtype=default_real_type

* ruff

* fix TODOs

* demo_lagrange_variants

* default_real_type

* fix

* default_real_type in more demos

* dtype=default_real_type in tests

* Simplify constructors

* Simplify

* add comments

* Revert "Simplify constructors" - experiment to see if this caused failing test

This reverts commit affecd2d0fc36be8c2732debe38222b017f1e58b.

* consts

* Revert "consts"

This reverts commit d697a21e09d1808a623d923f77d5e06a010a846d.

* Revert "Revert "Simplify constructors" - experiment to see if this caused failing test"

This reverts commit 98e377758941d029f5cf3cda796b9b8cbb40b447.

* Small fix

* move implementation to cpp file

* Reordering and linting

* Simplifications

---------

+++ python/test/unit/fem/test_complex_assembler.py
@@ -24,7 +24,7 @@ def test_complex_assembly(complex_dtype):
 
     real_dtype = np.real(complex_dtype(1.0)).dtype
     mesh = create_unit_square(MPI.COMM_WORLD, 10, 10, dtype=real_dtype)
-    P2 = element("Lagrange", mesh.basix_cell(), 2)
+    P2 = element("Lagrange", mesh.basix_cell(), 2, dtype=real_dtype)
     V = functionspace(mesh, P2)
     u = ufl.TrialFunction(V)
     v = ufl.TestFunction(V)
@@ -78,7 +78,7 @@ def test_complex_assembly_solve(complex_dtype, cg_solver):
     degree = 3
     real_dtype = np.real(complex_dtype(1.0)).dtype
     mesh = create_unit_square(MPI.COMM_WORLD, 20, 20, dtype=real_dtype)
-    P = element("Lagrange", mesh.basix_cell(), degree)
+    P = element("Lagrange", mesh.basix_cell(), degree, dtype=real_dtype)
     V = functionspace(mesh, P)
 
     x = ufl.SpatialCoordinate(mesh)
